OSM infrastructure data used in creation of transport infrastructure data packs.

Contains all columns required for IM function, oi_recode_max_speed, oi_recode_road_, oi_active_cycle, oi_is_lit.

example_data

Format

A data frame with 29,911 features described by 29 variables.

Source

generate_example_data.R

Examples

library(sf)
#> Linking to GEOS 3.8.0, GDAL 3.0.4, PROJ 6.3.1; sf_use_s2() is TRUE
example_data   # Lazy Loading. Data becomes visible as soon as it is requested. 
#> Simple feature collection with 29911 features and 28 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -1.889658 ymin: 53.71387 xmax: -1.285435 ymax: 53.87625
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>      osm_id               name      highway waterway aerialway barrier man_made
#> 17  1709451        Armley Road unclassified     <NA>      <NA>    <NA>     <NA>
#> 18  1709456        Hanover Way     tertiary     <NA>      <NA>    <NA>     <NA>
#> 19  1709460         Hyde Place  residential     <NA>      <NA>    <NA>     <NA>
#> 113 2390554    Armley Gyratory        trunk     <NA>      <NA>    <NA>     <NA>
#> 114 2447290        Park Avenue  residential     <NA>      <NA>    <NA>     <NA>
#> 115 2447535 Ingram Distributor        trunk     <NA>      <NA>    <NA>     <NA>
#> 116 2526748       Hunslet Road primary_link     <NA>      <NA>    <NA>     <NA>
#> 117 2526750               <NA>   trunk_link     <NA>      <NA>    <NA>     <NA>
#> 129 2956848         Manor View  residential     <NA>      <NA>    <NA>     <NA>
#> 130 2956850       Manor Avenue  residential     <NA>      <NA>    <NA>     <NA>
#>     foot bicycle access service maxspeed oneway kerb footway sidewalk cycleway
#> 17  <NA>    <NA>   <NA>    <NA>     <NA>    yes <NA>    <NA>     <NA>     <NA>
#> 18  <NA>    <NA>   <NA>    <NA>   30 mph   <NA> <NA>    <NA>     <NA>     <NA>
#> 19  <NA>    <NA>   <NA>    <NA>   20 mph   <NA> <NA>    <NA>     <NA>     <NA>
#> 113  yes     yes   <NA>    <NA>   40 mph    yes <NA>    <NA>     <NA>     <NA>
#> 114  yes     yes   <NA>    <NA>     <NA>   <NA> <NA>    <NA>     <NA>     <NA>
#> 115  yes     yes   <NA>    <NA>   40 mph    yes <NA>    <NA>       no     <NA>
#> 116 <NA>    <NA>   <NA>    <NA>   30 mph    yes <NA>    <NA>     <NA>     <NA>
#> 117 <NA>    <NA>   <NA>    <NA>     <NA>    yes <NA>    <NA>     <NA>     <NA>
#> 129 <NA>    <NA>   <NA>    <NA>     <NA>   <NA> <NA>    <NA>     <NA>     <NA>
#> 130 <NA>    <NA>   <NA>    <NA>     <NA>   <NA> <NA>    <NA>     <NA>     <NA>
#>     segregated crossing  lit tactile_paving surface smoothness width est_width
#> 17        <NA>     <NA> <NA>           <NA>    <NA>       <NA>  <NA>      <NA>
#> 18        <NA>     <NA>  yes           <NA>    <NA>       <NA>  <NA>      <NA>
#> 19        <NA>     <NA> <NA>           <NA>    <NA>       <NA>  <NA>      <NA>
#> 113       <NA>     <NA>  yes           <NA>    <NA>       <NA>  <NA>      <NA>
#> 114       <NA>     <NA> <NA>           <NA> asphalt       <NA>  <NA>      <NA>
#> 115       <NA>     <NA>  yes           <NA>    <NA>       <NA>  <NA>      <NA>
#> 116       <NA>     <NA>  yes           <NA>    <NA>       <NA>  <NA>      <NA>
#> 117       <NA>     <NA> <NA>           <NA>    <NA>       <NA>  <NA>      <NA>
#> 129       <NA>     <NA> <NA>           <NA>    <NA>       <NA>  <NA>      <NA>
#> 130       <NA>     <NA> <NA>           <NA>    <NA>       <NA>  <NA>      <NA>
#>     lit_by_led z_order
#> 17        <NA>       3
#> 18        <NA>       4
#> 19        <NA>       3
#> 113       <NA>       8
#> 114       <NA>       3
#> 115       <NA>       8
#> 116       <NA>       7
#> 117       <NA>       8
#> 129       <NA>       3
#> 130       <NA>       3
#>                                                                                                                                                                                                                    other_tags
#> 17                                                                                                                                                                                                               "lanes"=>"1"
#> 18                                                                                                                                                                                                                       <NA>
#> 19                                                                                                                                                                                                                       <NA>
#> 113                                                                                                  "horse"=>"yes","junction"=>"roundabout","lanes"=>"3","motorcar"=>"yes","ref"=>"A58","traffic_restriction"=>"no_stopping"
#> 114                                                                                                                                                                                          "horse"=>"yes","motorcar"=>"yes"
#> 115 "horse"=>"yes","lanes"=>"2","motorcar"=>"yes","old_name"=>"Gelderd Road","ref"=>"A58","source:ref"=>"photograph","source_ref:ref"=>"http://www.flickr.com/photos/pobice/3336544566/","traffic_restriction"=>"no_stopping"
#> 116                                                                                                                                                                                  "lanes"=>"2","turn:lanes"=>"right|right"
#> 117                                                                                                                                                                                                                      <NA>
#> 129                                                                                                                                                                                                                      <NA>
#> 130                                                                                                                                                                                                                      <NA>
#>                           geometry
#> 17  LINESTRING (-1.574397 53.79...
#> 18  LINESTRING (-1.55754 53.799...
#> 19  LINESTRING (-1.558721 53.80...
#> 113 LINESTRING (-1.5688 53.7942...
#> 114 LINESTRING (-1.497017 53.83...
#> 115 LINESTRING (-1.566073 53.79...
#> 116 LINESTRING (-1.532446 53.78...
#> 117 LINESTRING (-1.525161 53.78...
#> 129 LINESTRING (-1.573466 53.81...
#> 130 LINESTRING (-1.573725 53.81...