OSM points data used in the creation of transport infrastructure data packs.

OSM points data used in the creation of transport infrastructure data packs.

example_data_pois

Format

A dat frame of OSM ways.

Source

generate_example_data.R

Examples

library(sf)
example_data_pois
#> Simple feature collection with 9192 features and 30 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -1.587359 ymin: 53.77851 xmax: -1.510865 ymax: 53.82309
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    osm_id                      name barrier           highway  ref address
#> 1  350950   Elland Road Interchange    <NA> motorway_junction    2    <NA>
#> 2  350955 Cemetery Road Interchange    <NA> motorway_junction   2a    <NA>
#> 3  350966       Holbeck Interchange    <NA> motorway_junction    3    <NA>
#> 4  358309                     Leeds    <NA>              <NA> <NA>    <NA>
#> 5  414438                      <NA>    <NA>          crossing <NA>    <NA>
#> 6  418000                      <NA>    <NA>              <NA> <NA>    <NA>
#> 7  418047                      <NA>    <NA>              <NA> <NA>    <NA>
#> 8  611056                      <NA>    <NA>   traffic_signals <NA>    <NA>
#> 9  643805                      <NA>    <NA>   traffic_signals <NA>    <NA>
#> 10 643835                      <NA>    <NA>   traffic_signals <NA>    <NA>
#>                                                    is_in place man_made foot
#> 1                                                   <NA>  <NA>     <NA> <NA>
#> 2                                                   <NA>  <NA>     <NA> <NA>
#> 3                                                   <NA>  <NA>     <NA> <NA>
#> 4  UK,United Kingdom, Yorkshire,West Yorkshire, Airedale  city     <NA> <NA>
#> 5                                                   <NA>  <NA>     <NA> <NA>
#> 6                                                   <NA>  <NA>     <NA> <NA>
#> 7                                                   <NA>  <NA>     <NA> <NA>
#> 8                                                   <NA>  <NA>     <NA> <NA>
#> 9                                                   <NA>  <NA>     <NA> <NA>
#> 10                                                  <NA>  <NA>     <NA> <NA>
#>    bicycle access service maxspeed oneway    kerb footway sidewalk cycleway
#> 1     <NA>   <NA>    <NA>     <NA>   <NA>    <NA>    <NA>     <NA>     <NA>
#> 2     <NA>   <NA>    <NA>     <NA>   <NA>    <NA>    <NA>     <NA>     <NA>
#> 3     <NA>   <NA>    <NA>     <NA>   <NA>    <NA>    <NA>     <NA>     <NA>
#> 4     <NA>   <NA>    <NA>     <NA>   <NA>    <NA>    <NA>     <NA>     <NA>
#> 5     <NA>   <NA>    <NA>     <NA>   <NA>    <NA>    <NA>     <NA>     <NA>
#> 6     <NA>   <NA>    <NA>     <NA>   <NA>    <NA>    <NA>     <NA>     <NA>
#> 7     <NA>   <NA>    <NA>     <NA>   <NA>    <NA>    <NA>     <NA>     <NA>
#> 8     <NA>   <NA>    <NA>     <NA>   <NA> lowered    <NA>     <NA>     <NA>
#> 9     <NA>   <NA>    <NA>     <NA>   <NA>    <NA>    <NA>     <NA>     <NA>
#> 10    <NA>   <NA>    <NA>     <NA>   <NA>    <NA>    <NA>     <NA>     <NA>
#>    segregated        crossing  lit tactile_paving surface smoothness width
#> 1        <NA>            <NA> <NA>           <NA>    <NA>       <NA>  <NA>
#> 2        <NA>            <NA> <NA>           <NA>    <NA>       <NA>  <NA>
#> 3        <NA>            <NA> <NA>           <NA>    <NA>       <NA>  <NA>
#> 4        <NA>            <NA> <NA>           <NA>    <NA>       <NA>  <NA>
#> 5        <NA>            <NA> <NA>           <NA>    <NA>       <NA>  <NA>
#> 6        <NA>            <NA> <NA>           <NA>    <NA>       <NA>  <NA>
#> 7        <NA>            <NA> <NA>           <NA>    <NA>       <NA>  <NA>
#> 8        <NA> traffic_signals <NA>            yes    <NA>       <NA>  <NA>
#> 9        <NA>            <NA> <NA>           <NA>    <NA>       <NA>  <NA>
#> 10       <NA>            <NA> <NA>           <NA>    <NA>       <NA>  <NA>
#>    est_width lit_by_led amenity
#> 1       <NA>       <NA>    <NA>
#> 2       <NA>       <NA>    <NA>
#> 3       <NA>       <NA>    <NA>
#> 4       <NA>       <NA>    <NA>
#> 5       <NA>       <NA>    <NA>
#> 6       <NA>       <NA>    <NA>
#> 7       <NA>       <NA>    <NA>
#> 8       <NA>       <NA>    <NA>
#> 9       <NA>       <NA>    <NA>
#> 10      <NA>       <NA>    <NA>
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            other_tags
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                <NA>
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                <NA>
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                <NA>
#> 4  "alt_name:la"=>"Leodis;Loidis","is_capital"=>"county","is_in:continent"=>"Europe","is_in:country"=>"United Kingdom","is_in:country_code"=>"GB","is_in:county"=>"West Yorkshire","is_in:historic_county"=>"Yorkshire","name:af"=>"Leeds","name:als"=>"Leeds","name:am"=>"ሊድስ","name:ang"=>"Loidis","name:ar"=>"ليدز","name:arc"=>"ܠܝܕܣ","name:arz"=>"ليدز","name:ast"=>"Leeds","name:az"=>"Lids","name:azb"=>"لیدز","name:be"=>"Лідс","name:be-x-old"=>"Лідс","name:bg"=>"Лийдс","name:bn"=>"লিডস","name:br"=>"Leeds","name:bs"=>"Leeds","name:ca"=>"Leeds","name:ce"=>"Лидс","name:ceb"=>"Leeds)","name:ckb"=>"لیدز","name:cs"=>"Leeds","name:cy"=>"Leeds","name:da"=>"Leeds","name:de"=>"Leeds","name:diq"=>"Leeds","name:el"=>"Ληντς","name:en"=>"Leeds","name:eo"=>"Leeds","name:es"=>"Leeds","name:et"=>"Leeds","name:eu"=>"Leeds","name:fa"=>"لیدز","name:fi"=>"Leeds","name:fo"=>"Leeds","name:fr"=>"Leeds","name:frr"=>"Leeds","name:fy"=>"Leeds","name:ga"=>"Leeds","name:gl"=>"Leeds","name:gu"=>"લીડ્ઝ","name:gv"=>"Leeds","name:ha"=>"Leeds","name:he"=>"לידס","name:hi"=>"लीड्स","name:hr"=>"Leeds","name:hsb"=>"Leeds","name:hu"=>"Leeds","name:hy"=>"Լիդս","name:ia"=>"Leeds","name:id"=>"Leeds","name:ie"=>"Leeds","name:io"=>"Leeds","name:is"=>"Leeds","name:it"=>"Leeds","name:ja"=>"リーズ","name:ka"=>"ლიდზი","name:kn"=>"ಲೀಡ್ಸ್","name:ko"=>"리즈","name:ku"=>"Leeds","name:ky"=>"Лидс","name:la"=>"Ledesia","name:lmo"=>"Leeds","name:lt"=>"Lidsas","name:lv"=>"Līdsa","name:mk"=>"Лидс","name:ml"=>"ലീഡ്സ്","name:mni"=>"ꯂꯤꯗ꯭ꯁ","name:mr"=>"लीड्स","name:mt"=>"Leeds","name:mzn"=>"لیدز","name:nan"=>"Leeds","name:nds"=>"Leeds","name:nl"=>"Leeds","name:nn"=>"Leeds","name:no"=>"Leeds","name:oc"=>"Leeds","name:os"=>"Лидс","name:pa"=>"ਲੀਡਸ","name:pl"=>"Leeds","name:pms"=>"Leeds","name:pnb"=>"لیڈز","name:pt"=>"Leeds","name:qu"=>"Leeds","name:ro"=>"Leeds","name:ru"=>"Лидс","name:sc"=>"Leeds","name:sco"=>"Leeds","name:sh"=>"Leeds","name:sk"=>"Leeds","name:sl"=>"Leeds","name:sr"=>"Лидс","name:sv"=>"Leeds","name:sw"=>"Leeds","name:szl"=>"Leeds","name:ta"=>"லீட்சு","name:te"=>"లీడ్స్","name:tg"=>"Лидс","name:tl"=>"Leeds","name:tr"=>"Leeds","name:tt"=>"Лидс","name:tw"=>"Leeds","name:ug"=>"Ritis","name:uk"=>"Лідс","name:ur"=>"لیڈز","name:uz"=>"Lids","name:vep"=>"Lids","name:vi"=>"Leeds","name:vo"=>"Leeds","name:war"=>"Leeds","name:wuu"=>"利兹","name:yi"=>"לידז","name:yue"=>"列斯","name:zh"=>"利兹","name:zh-Hans"=>"利兹","name:zh-Hant"=>"里茲","population"=>"766399","wikidata"=>"Q39121","wikipedia"=>"en:Leeds"
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                <NA>
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 "railway"=>"switch"
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 "railway"=>"switch"
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           "crossing_ref"=>"pelican"
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              "traffic_signals:direction"=>"forward"
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 "traffic_signals"=>"signal","traffic_signals:direction"=>"forward"
#>                      geometry
#> 1  POINT (-1.563325 53.78022)
#> 2   POINT (-1.55534 53.78336)
#> 3  POINT (-1.552644 53.78419)
#> 4  POINT (-1.543794 53.79742)
#> 5  POINT (-1.545744 53.79519)
#> 6  POINT (-1.515634 53.79292)
#> 7  POINT (-1.568322 53.79144)
#> 8  POINT (-1.533996 53.79456)
#> 9   POINT (-1.53094 53.78382)
#> 10 POINT (-1.535514 53.78833)