Function to recategorise OSM infrastrcutre based on pedestrian walkability.

Determins whether or not a piece of OSM infrastructure is walkable or not depending on whether the feature is included in the default walking network defined by osmextract

oi_active_walk(osm_sf, remove=FALSE)

Arguments

osm_sf

A sf and data.frame object containing OpenStreetMap infrastructure data, obtained from the osmextract function.

remove

Boolean - If TRUE, features that do not form part of a walkable network are removed from the osm_sf before being returned

Value

an sf object with oi_walk column added.

Details

This function mimics the walking filtering steps applied by the oe_get_network function, part of the osmextract package.

Note: the osm_sf must contain the following tags: c("highway", "foot", "access", "service")

Examples

data = example_data
example_output = oi_active_walk(data)