Function to recategorise OSM data based on compliance with the Inclusive Mobility (IM)

This function is based on the Inclusive Mobility (IM) guide UK.

oi_inclusive_mobility(osm_sf)

Arguments

osm_sf

A sf and data.frame object containing OpenStreetMap infrastructure data, obtained from the osmextract function.

Value

The osm_sf data frame is returned with additional inclusive mobility columns returned.

Details

See details in the link above, the function's source code, and our justification vignette.

Note: the osm_sf must contain the following tags: c("cycleway", "bicycle", "wheelchair", "kerb", "disabled", "mobility_scooter", "foot", "handicap", "lit", "access", "sidewalk", "footway", "incline", "width", "smoothness", "est_width", "sidewalk_left", "ramp_wheelchair", "priority", "footway_left", "sidewalk_both_surface", "footway_surface", "pedestrian", "footway_right", "ramp", "sidewalk_right", "sidewalk_both_width", "path", "sidewalk_left_width", "sidewalk_right_width", "sidewalk_right_surface", "sidewalk_left_surface", "maxspeed", "segregated", "sloped_curb", "surface", "tactile_paving")

Examples

library(sf)
internal_data = example_data
dim(internal_data)
#> [1] 13236    35
output = oi_inclusive_mobility(internal_data)