OSM infrastructure data to showcase the recode_road_class openinfra function.

OSM infrastructure data to showcase the recode_road_class openinfra function.

example_data

Format

A data frame of infrastructure features for Leeds, with 36,065 features and 16 variable. Each feature is described by the following columns.

osm_id

Unique feature type (Node|Way|Relation) and numeric ID combination. e.g OSM Way 2746 has osm_id = W2746.

name

This key is set to the primary name of the feature in the real world.

highway

Main key used to identify any kind of road, street or path.

waterway

Key used to describe water flows like rivers, streams, boat-yards. Or, elements that interact with water like dams, weirs, locks etc.

aerialway

Key used for various forms of passenger or goods transport that uses wires, including cable-cars, chair-lifts, gondolas etc.

barrier

Key used for a physical ground structures which blocks or impeads movement.

man_made

Key used to identify man made structures.

access

Key to describe restrictions on the use of highways, and other trasnportation routes (warterways|railways|aerialways) as well as facilities such as buildings.

service

Key to show feature is used for access to a building, service station, beach, campsite, industrial estate, business park etc. Commonly also used for acces to parking, driveways and alleyways.

bicycle

Key to indicate legal restrictions for cyclists. i.e "dismount" implies dismount the cycle, "yes" implies suitable for cyclists.

foot

Key to indicate legal resrictions for pedestrians

oneway

Key to indicate the oneway access restriction on highways, and other ways, for vehicles (car|bicycle|truck etc.) as appropriate

maxspeed

Key used to define the maximum legal speed limit for general traffic on a particular road.

z_order

A field in osm2pgsql datamodel that map OSM Layers in OSM tags to a number in osm2pgsql model. See Within Renderers

other_tags

Any additional tags attributed to a feature that is not returned as its own column (muhc like this column) by default.

geometry

A column that contains the geometry of the feature in question. Typically either Linestring, MultiLinestring, Polygon, MultiPolygon.

place_name = "Leeds"
provider = "bbbike"

total_place = osmextract::oe_get(
place = place_name,
provider = provider,
layer = "lines",
never_skip_vectortranslate = TRUE,
force_download = TRUE,
quiet = FALSE,
extra_tags = c(all_extra_tags, "oneway", "maxspeed") # Add in "oneway" for recode_road_class)

Source

Created using the osmextract package for R using oe_get() - params shown above.

Examples

example_data   # Lazy Loading. Data becomes visible as soon as it is requested. 
#> Error in eval(expr, envir, enclos): object 'example_data' not found